Add files via upload
Added fixes for the animal to be destroyed when hitting boy, added boy.setCollider so that animal should touch boy and not gun when life time reduces. created a separate function for the shooting of bullets. fixed final game over font and text font. bullet position and boy position updated. updated jungle image to have the image copied side by side so that no need to do scaling and yet have the infinite ground image.